routes: migrate verifications domain (2 internal routes)
Move to src/server/routes/verifications.ts:
- POST /api/internal/verification-result
- POST /api/internal/user-question/submit

Move askSubmittedToolUseIds dedup Set into the new module.
Drop now-unused imports (validateAnswers, crossValidate, UserQuestion,
buildAskResponseEnvelope, findAskResponseAnswers, inlineFileImages).

Net: server.ts -167 lines (3000 → 2833). No behaviour change.
