+++ src/server/routes/dispatcher.ts
@@ -40,6 +40,7 @@ import { sessionsGitRoutes } from "./sessions-git.js";
 import { previewRoutes } from "./preview.js";
 import { sessionsContentRoutes } from "./sessions-content.js";
 import { sessionsProposalsRoutes } from "./sessions-proposals.js";
+import { verificationsRoutes } from "./verifications.js";
 
 // Registered route arrays — populated as domain migrations land. Each
 // `routes/<domain>.ts` exports a `<domain>Routes: Route[]` array. Until all
@@ -71,6 +72,7 @@ const allRoutes: ReadonlyArray<Route> = [
 	...previewRoutes,
 	...sessionsContentRoutes,
 	...sessionsProposalsRoutes,
+	...verificationsRoutes,
 ];
 
 /**
