+++ src/server/server.ts
@@ -34,17 +34,6 @@ import { BgProcessManager } from "./agent/bg-process-manager.js";
 
 import { isGitRepo, getRepoRoot, stripTokenFromGitUrl, detectPrimaryBranch } from "./skills/git.js";
 import { VerificationHarness } from "./agent/verification-harness.js";
-import { validateAnswers, crossValidate, type UserQuestion } from "./agent/ask-user-choices-validation.js";
-import { buildAskResponseEnvelope, findAskResponseAnswers } from "../shared/ask-envelope.js";
-
-// In-memory dedup guard for ask_user_choices /submit. Keyed by
-// `${sessionId}::${toolUseId}`. Populated synchronously before enqueuing the
-// response envelope so a concurrent duplicate /submit returns alreadySubmitted
-// even when the transcript hasn't yet reflected the first envelope.
-// Entries are also refilled from the transcript check, so survive process
-// restarts via the transcript fallback in findAskResponseAnswers.
-const askSubmittedToolUseIds = new Set<string>();
-import { inlineFileImages } from "./agent/inline-file-images.js";
 import { StaffManager } from "./agent/staff-manager.js";
 import { TriggerEngine } from "./agent/staff-trigger-engine.js";
 import { PreferencesStore } from "./agent/preferences-store.js";
@@ -2834,162 +2823,6 @@ async function handleApiRoute(
 			}
 		}
 
-		json({ ok: true });
-		return;
-	}
-
-	// POST /api/internal/verification-result
-	if (url.pathname === "/api/internal/verification-result" && req.method === "POST") {
-		const body = await readBody(req);
-		if (!body?.sessionId || !body?.verdict || !body?.summary || typeof body.sessionId !== "string" || typeof body.verdict !== "string" || typeof body.summary !== "string") {
-			json({ error: "Missing required fields: sessionId, verdict, summary" }, 400);
-			return;
-		}
-		const resolver = verificationHarness.pendingResults.get(body.sessionId);
-		if (!resolver) {
-			json({ error: "No pending verification for this session" }, 404);
-			return;
-		}
-		// Support report_html_file: server reads file directly (avoids tool output limits for large reports)
-		if (typeof body.report_html === "string" && typeof body.report_html_file === "string") {
-			json({ error: "Provide either report_html or report_html_file, not both" }, 400);
-			return;
-		}
-		let reportHtml: string | undefined = typeof body.report_html === "string" ? body.report_html : undefined;
-		if (!reportHtml && typeof body.report_html_file === "string") {
-			try {
-				let filePath = body.report_html_file;
-				// Resolve relative paths against the session's CWD
-				if (!path.isAbsolute(filePath)) {
-					const session = sessionManager.getSession(body.sessionId);
-					if (session) filePath = path.resolve(session.cwd, filePath);
-				}
-				// On Windows, POSIX paths from Git Bash (/tmp/...) resolve to C:\tmp\... which doesn't exist.
-				// Fall back to the system TEMP directory for /tmp/ paths.
-				if (process.platform === "win32" && !fs.existsSync(filePath) && body.report_html_file.startsWith("/tmp/")) {
-					const tempDir = process.env.TEMP || process.env.TMP || "C:\\Windows\\Temp";
-					const tempResolved = path.join(tempDir, body.report_html_file.slice(5));
-					if (fs.existsSync(tempResolved)) filePath = tempResolved;
-				}
-				const stat = fs.statSync(filePath);
-				const MAX_REPORT_SIZE = 10 * 1024 * 1024; // 10 MB
-				if (stat.size > MAX_REPORT_SIZE) {
-					json({ error: `Report file too large (${stat.size} bytes, max ${MAX_REPORT_SIZE})` }, 400);
-					return;
-				}
-				reportHtml = fs.readFileSync(filePath, "utf-8");
-			} catch (e: any) {
-				json({ error: `Failed to read report file: ${e.message}` }, 400);
-				return;
-			}
-		}
-		// Inline any <img src="file://..."> references so the report renders from
-		// the browser's blob origin (cross-origin file:// loads are blocked).
-		if (reportHtml) {
-			const session = sessionManager.getSession(body.sessionId);
-			if (session?.cwd) {
-				try {
-					reportHtml = inlineFileImages(reportHtml, session.cwd, {
-						logger: (msg) => console.warn(msg),
-					});
-				} catch (err: any) {
-					console.warn(`[verification] inlineFileImages failed: ${err?.message || err}`);
-				}
-			}
-		}
-		resolver({
-			verdict: body.verdict === "pass",
-			summary: body.summary,
-			reportHtml,
-		});
-		json({ ok: true });
-		return;
-	}
-
-	// POST /api/internal/user-question/submit  — called by the UI widget with answers.
-	// Non-blocking model: appends a tagged user message to the session transcript
-	// via `enqueuePrompt` (the normal user-prompt path), which persists to .jsonl,
-	// broadcasts, and wakes the agent. Idempotent: duplicate submits for the same
-	// tool_use_id are a no-op (the second tab's submit is swallowed).
-	// See src/shared/ask-envelope.ts for the envelope format.
-	if (url.pathname === "/api/internal/user-question/submit" && req.method === "POST") {
-		const body = await readBody(req);
-		const { sessionId, toolUseId, answers } = body || {};
-		if (typeof sessionId !== "string" || typeof toolUseId !== "string" || !Array.isArray(answers)) {
-			json({ error: "Missing required fields: sessionId, toolUseId, answers" }, 400);
-			return;
-		}
-		const answerErr = validateAnswers(answers);
-		if (answerErr) { json({ error: answerErr }, 400); return; }
-		const session = sessionManager.getSession(sessionId);
-		if (!session) { json({ error: "Unknown session" }, 404); return; }
-
-		// Pull the transcript to locate the original tool_use (for cross-validation)
-		// and to detect duplicate submits (multi-tab / network retry).
-		let messages: any[] = [];
-		try {
-			const msgsResp = await session.rpcClient.getMessages();
-			const raw = msgsResp?.data?.messages || msgsResp?.data;
-			if (Array.isArray(raw)) messages = raw;
-		} catch (e: any) {
-			json({ error: `Could not load transcript: ${e?.message || String(e)}` }, 500);
-			return;
-		}
-
-		// Idempotency: if a response envelope for this toolUseId already exists,
-		// return success without appending again. Check in-memory guard first
-		// (covers the race where a duplicate /submit arrives before the first
-		// envelope has propagated into the transcript), then the transcript
-		// (covers process restart / external writers).
-		const dedupKey = `${sessionId}::${toolUseId}`;
-		if (askSubmittedToolUseIds.has(dedupKey)) {
-			json({ ok: true, alreadySubmitted: true });
-			return;
-		}
-		const existing = findAskResponseAnswers(messages, toolUseId);
-		if (existing) {
-			askSubmittedToolUseIds.add(dedupKey);
-			json({ ok: true, alreadySubmitted: true });
-			return;
-		}
-
-		// Locate the ask_user_choices tool_use block; use its input to cross-validate.
-		let matchedQuestions: UserQuestion[] | null = null;
-		for (const m of messages) {
-			if (!m || m.role !== "assistant" || !Array.isArray(m.content)) continue;
-			for (const b of m.content) {
-				if (!b) continue;
-				const isToolUse = b.type === "toolCall" || b.type === "tool_use";
-				if (!isToolUse) continue;
-				if (b.name !== "ask_user_choices") continue;
-				if (b.id !== toolUseId) continue;
-				const args = b.arguments ?? b.input;
-				if (args && Array.isArray(args.questions)) {
-					matchedQuestions = args.questions as UserQuestion[];
-				}
-				break;
-			}
-			if (matchedQuestions) break;
-		}
-		if (!matchedQuestions) {
-			json({ error: "No matching ask_user_choices tool call in transcript" }, 404);
-			return;
-		}
-		const crossErr = crossValidate(matchedQuestions, answers);
-		if (crossErr) { json({ error: crossErr }, 400); return; }
-
-		const envelope = buildAskResponseEnvelope(toolUseId, answers);
-		// Mark as submitted BEFORE awaiting enqueuePrompt so a concurrent
-		// duplicate /submit is rejected deterministically.
-		askSubmittedToolUseIds.add(dedupKey);
-		try {
-			await sessionManager.enqueuePrompt(sessionId, envelope);
-		} catch (e: any) {
-			// Roll back the dedup flag so the caller can retry.
-			askSubmittedToolUseIds.delete(dedupKey);
-			json({ error: `Failed to enqueue response: ${e?.message || String(e)}` }, 500);
-			return;
-		}
 		json({ ok: true });
 		return;
 	}
